segformer_polygons/requirements.txt
torch
transformers>=4.42
accelerate>=0.26.0
numpy
Pillow
scikit-image
scipy
tqdm
